+++ components/src/NavBar/MenuLink.js
@@ -1,10 +1,12 @@
 import styled from "styled-components";
+import { themeGet } from "styled-system";
 import PropTypes from "prop-types";
-import { Link } from "../Link";
 import theme from "../theme";
 
-const MenuLink = styled(Link)({
+const MenuLink = styled.a(({ color, hoverColor, hoverBackground }) => ({
   display: "block",
+  color: themeGet(`colors.${color}`, color)(color),
+  textDecoration: "none",
   border: "none",
   backgroundColor: "transparent",
   lineHeight: theme.lineHeights.base,
@@ -14,21 +16,25 @@ const MenuLink = styled(Link)({
   borderRadius: theme.radii.medium,
   "&:hover, &:focus": {
     outline: "none",
-    backgroundColor: theme.colors.black
+    color: themeGet(`colors.${hoverColor}`, hoverColor)(hoverColor),
+    backgroundColor: themeGet(`colors.${hoverBackground}`, hoverBackground)(hoverBackground),
+    cursor: "pointer"
   },
   "&:disabled": {
     opacity: ".5"
   }
-});
+}));
 
 MenuLink.propTypes = {
   color: PropTypes.string,
-  underline: PropTypes.bool
+  hoverColor: PropTypes.string,
+  hoverBackground: PropTypes.string
 };
 
 MenuLink.defaultProps = {
   color: theme.colors.white,
-  underline: false
+  hoverColor: theme.colors.lightBlue,
+  hoverBackground: theme.colors.black
 };
 
 export default MenuLink;