+++ components/src/NavBar/NavBar.story.js
@@ -125,4 +125,7 @@ storiesOf("NavBar", module)
   .add("With custom link components", () => <NavBar menuData={{ primaryMenu: primaryMenuCustomLinks }} />)
   .add("With subtext", () => (
     <NavBar subtext="Logo Subtext" menuData={{ primaryMenu: primaryMenuCustomLinks, search }} />
+  ))
+  .add("With alternate themeColor", () => (
+    <NavBar subtext="Logo Subtext" menuData={{ primaryMenu: primaryMenuCustomLinks }} themeColor="white" />
   ));